1.7.0 release initial commit

+++ docker/build/contraxsuite-app/start.sh
@@ -1,7 +1,6 @@
 #!/bin/bash
 
 set -e
-source /webdav_upload.sh
 
 PROJECT_DIR="/contraxsuite_services"
 VENV_PATH="/contraxsuite_services/venv/bin/activate"
@@ -37,7 +36,7 @@ fi
 
 if [ "$1" == "save-dump" ]; then
 
-su - ${SHARED_USER_NAME} -c "${ACTIVATE_VENV} && \
+su ${SHARED_USER_NAME} -c "${ACTIVATE_VENV} && \
     python manage.py force_migrate common && \
     python manage.py force_migrate && \
     python manage.py dump_data --dst-file=fixtures/additional/app-dump.json \
@@ -45,25 +44,52 @@ su - ${SHARED_USER_NAME} -c "${ACTIVATE_VENV} && \
 
 elif [ "$1" == "daphne" ]; then
 
-    echo "Preparing jqwidgets..."
-    JQWIDGETS_ZIP=/third_party_dependencies/$(basename ${DOCKER_DJANGO_JQWIDGETS_ARCHIVE})
-    VENDOR_DIR=/static/vendor
-    rm -rf ${VENDOR_DIR}/jqwidgets
-    unzip ${JQWIDGETS_ZIP} "jqwidgets/*" -d ${VENDOR_DIR}
+    if [[ -f "/static/vendor/jqwidgets/jqx-all.js" ]]; then
+        echo "JQWidgets are bundled within the Contraxsuite Docker image. Not running collectstatic."
+    else
+        echo "JQWidgets are not bundled within the Docker Image. Checking the volume..."
 
-    echo "Updating customizable notification templates in media folder..."
+        JQWIDGETS_ZIP="/third_party_dependencies/jqwidgets.zip"
+        if [[ -f "${JQWIDGETS_ZIP}" ]]; then
+            echo "Using jqwidgets: ${JQWIDGETS_ZIP}..."
+            VENDOR_DIR=/static/vendor
+            rm -rf ${VENDOR_DIR}/jqwidgets
+            unzip ${JQWIDGETS_ZIP} "jqwidgets/*" -d ${VENDOR_DIR}
+
+            su ${SHARED_USER_NAME} -c "${ACTIVATE_VENV} && python manage.py collectstatic --noinput"
+        else
+            echo "Can't find JQWidgets neither in the Docker image nor at ${JQWIDGETS_ZIP}."
+            exit 1
+        fi
+    fi
+
+
     cat /build.info > /contraxsuite_services/staticfiles/version.txt
     echo "" >> /contraxsuite_services/staticfiles/version.txt
     cat /build.uuid >> /contraxsuite_services/staticfiles/version.txt
 
+    if [[ -d "/contraxsuite_frontend" && -d "/contraxsuite_frontend_nginx_volume" ]]; then
+        echo "Copying embedded frontend files into the shared Nginx volume..."
+        echo "Embedded frontend version:"
+        cat /contraxsuite_frontend/versionFE.txt
+        echo "Embedded frontend build info:"
+        cat /contraxsuite_frontend/version.txt
+        rm -rf /contraxsuite_frontend_nginx_volume/*
+        echo "API_HOST=${DOCKER_DJANGO_HOST_NAME}" > /contraxsuite_frontend/.env
+        echo "WS_HOST=${DOCKER_DJANGO_HOST_NAME}" >> /contraxsuite_frontend/.env
+
+        cp -R /contraxsuite_frontend/. /contraxsuite_frontend_nginx_volume/
+    fi
+
     echo "Copying notification templates"
+    source /webdav_upload.sh
     upload_files_to_webpath /contraxsuite_services/apps/notifications/notification_templates notification_templates example_ || true
 
     echo "Ensuring Django superuser is created..."
 
 
 # Indentation makes sense here
-su - ${SHARED_USER_NAME} -c "${ACTIVATE_VENV} && \
+su ${SHARED_USER_NAME} -c "${ACTIVATE_VENV} && \
     python manage.py force_migrate common && \
     python manage.py force_migrate users && \
     python manage.py force_migrate && \
@@ -73,12 +99,10 @@ from apps.deployment.tasks import usage_stats
 Deployment.objects.get_or_create(pk=1)
 usage_stats.apply()
 \" && \
-    python manage.py collectstatic --noinput && \
     python manage.py set_site && \
-    python manage.py create_superuser --username=${DOCKER_DJANGO_ADMIN_NAME} --email=${DOCKER_DJANGO_ADMIN_EMAIL} --password=${DOCKER_DJANGO_ADMIN_PASSWORD} && \
+    python manage.py create_superuser --username=Administrator --email=${DOCKER_DJANGO_ADMIN_EMAIL} --password=${DOCKER_DJANGO_ADMIN_PASSWORD} && \
     python manage.py loadnewdata fixtures/common/*.json && \
     python manage.py loadnewdata fixtures/additional/*.json && \
-    python manage.py init_app_data --data-dir=/data/data_update --arch-files && \
     python manage.py init_app_data --data-dir=/contraxsuite_services/fixtures/demo --upload-dict-data-from-repository && \
     if [ -d fixtures/customer_project ]; then python manage.py loadnewdata fixtures/customer_project/*.json; fi \
 "
@@ -92,7 +116,7 @@ usage_stats.apply()
       echo ""
       echo ""
 
-      su - ${SHARED_USER_NAME} -c "${ACTIVATE_VENV} && \
+      su ${SHARED_USER_NAME} -c "${ACTIVATE_VENV} && \
           ulimit -n 65535 && \
           python manage.py check && \
           daphne -b 0.0.0.0 -p 3355 asgi:application"
@@ -111,7 +135,7 @@ elif [ "$1" == "jupyter" ]; then
 
     set +e
     cat ${JUPYTER_ADD_DEBIAN_REQ} | xargs -r apt-get -y -q install
-    su - ${SHARED_USER_NAME} -c "${ACTIVATE_VENV} && pip install -r ${JUPYTER_ADD_REQ}"
+    su ${SHARED_USER_NAME} -c "${ACTIVATE_VENV} && pip install -r ${JUPYTER_ADD_REQ}"
     set -e
 
     mkdir -p /contraxsuite_services/notebooks
@@ -122,59 +146,66 @@ elif [ "$1" == "jupyter" ]; then
     cat /contraxsuite_services/jupyter_notebook_config.py > /home/${SHARED_USER_NAME}/.jupyter/jupyter_notebook_config.py
     chown -R -v ${SHARED_USER_NAME}:${SHARED_USER_NAME} /home/${SHARED_USER_NAME}/.jupyter
 
-    su - ${SHARED_USER_NAME} -c "${ACTIVATE_VENV} && \
+    su ${SHARED_USER_NAME} -c "${ACTIVATE_VENV} && \
     python -c \"
 from notebook.auth import passwd
 with open('/home/${SHARED_USER_NAME}/.jupyter/jupyter_notebook_config.py', 'a') as myfile:
     myfile.write('\\nc.NotebookApp.password = \'' + passwd('${DOCKER_DJANGO_ADMIN_PASSWORD}') + '\'')
 \""
-    su - ${SHARED_USER_NAME} -c "${ACTIVATE_VENV} && \
+    su ${SHARED_USER_NAME} -c "${ACTIVATE_VENV} && \
         ulimit -n 65535 && \
         jupyter notebook --port=8888 --no-browser --ip=0.0.0.0"
 elif [ $1 == "flower" ]; then
     echo "Starting Flower..."
 
-    su - ${SHARED_USER_NAME} -c "${ACTIVATE_VENV} && \
+    su ${SHARED_USER_NAME} -c "${ACTIVATE_VENV} && \
         ulimit -n 65535 && \
         flower -A apps --port=5555 --address=0.0.0.0 --url_prefix=${DOCKER_FLOWER_BASE_PATH}"
+elif [ $1 == "mock-imanage" ]; then
+    echo "Starting Mock IManage Server..."
+
+    su ${SHARED_USER_NAME} -c "${ACTIVATE_VENV} && \
+        ulimit -n 65535 && \
+        cd apps/imanage_integration/debug/ && \
+        python mock_imanage_server.py 65534"
 elif [ $1 == "celery-beat" ]; then
     echo "Starting Celery Beat and Serial Tasks Worker..."
 
-    su - ${SHARED_USER_NAME} -c "${ACTIVATE_VENV} && \
+    su ${SHARED_USER_NAME} -c "${ACTIVATE_VENV} && \
         ulimit -n 65535 && \
         celery -A apps worker -B -Q serial --without-gossip --without-heartbeat --concurrency=1 -Ofair -n beat@%h --statedb=/data/celery_worker_state/worker.state --role=beat"
 elif [ $1 == "celery-high-prio" ]; then
     echo "Starting Celery High Priority Tasks Worker..."
 
-    su - ${SHARED_USER_NAME} -c "${ACTIVATE_VENV} && \
+    su ${SHARED_USER_NAME} -c "${ACTIVATE_VENV} && \
         ulimit -n 65535 && \
         celery -A apps worker -Q high_priority --without-gossip --without-heartbeat --concurrency=4 -Ofair -n high_priority@%h --statedb=/data/celery_worker_state/worker.state --role=high_prio"
 
 elif [ $1 == "celery-load" ]; then
     echo "Starting Celery Load Documents Tasks Worker..."
 
-    su - ${SHARED_USER_NAME} -c "${ACTIVATE_VENV} && \
+    su ${SHARED_USER_NAME} -c "${ACTIVATE_VENV} && \
         ulimit -n 65535 && \
         celery -A apps worker -Q doc_load --without-gossip --without-heartbeat --concurrency=1 -Ofair -n default_priority@%h --statedb=/data/celery_worker_state/worker.state --role=doc_load"
 
 elif [ $1 == "celery-master" ]; then
     echo "Starting Celery Master Low Resources Worker..."
 
-    su - ${SHARED_USER_NAME} -c "${ACTIVATE_VENV} && \
+    su ${SHARED_USER_NAME} -c "${ACTIVATE_VENV} && \
         ulimit -n 65535 && \
         celery -A apps worker -Q default,high_priority --without-gossip --without-heartbeat --concurrency=2 -Ofair -n master@%h --statedb=/data/celery_worker_state/worker.state --role=default_worker"
 elif [ $1 == "celery-single" ]; then
     echo "Starting Celery Default Priority Tasks Worker..."
 
-    su - ${SHARED_USER_NAME} -c "${ACTIVATE_VENV} && \
+    su ${SHARED_USER_NAME} -c "${ACTIVATE_VENV} && \
         ulimit -n 65535 && \
         python manage.py check && \
-        celery -A apps worker -Q default,high_priority --without-gossip --without-heartbeat --concurrency=${DOCKER_CELERY_CONCURRENCY} -Ofair -n default_priority@%h --statedb=/data/celery_worker_state/worker.state --role=default_worker"
+        celery -A apps worker -Q default,high_priority,worker_bcast --without-gossip --without-heartbeat --concurrency=${DOCKER_CELERY_CONCURRENCY} -Ofair -n default_priority@%h --statedb=/data/celery_worker_state/worker.state --role=default_worker"
 else
     echo "Starting Celery Default Priority Tasks Worker..."
 
-    su - ${SHARED_USER_NAME} -c "${ACTIVATE_VENV} && \
+    su ${SHARED_USER_NAME} -c "${ACTIVATE_VENV} && \
         ulimit -n 65535 && \
         python manage.py check && \
-        celery -A apps worker -Q default --without-gossip --without-heartbeat --concurrency=${DOCKER_CELERY_CONCURRENCY} -Ofair -n default_priority@%h --statedb=/data/celery_worker_state/worker.state --role=default_worker"
+        celery -A apps worker -Q default,worker_bcast --without-gossip --without-heartbeat --concurrency=${DOCKER_CELERY_CONCURRENCY} -Ofair -n default_priority@%h --statedb=/data/celery_worker_state/worker.state --role=default_worker"
 fi
